+++ ouroboros-network-framework/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Non-breaking changes
 
+* ghc-9.8 support.
+
 ## 0.10.2.0 -- 2023-12-14
 
 ### Non-breaking changes