+++ ouroboros-network-protocols/CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Breaking changes
 
+* Pipeline depth type changed from `Word32` to `Word16`.
+
 ### Non-breaking changes
+
+* ghc-9.8 support.
 
 ## 0.6.1.0 -- 2023-12-14
 
@@ -12,6 +16,7 @@
 
 * Testlib depends on `cardano-slotting`'s `testlib` at version
   `0.1.2.0` and uses its instances.
+
 * Use `io-sim-1.3.1.0`.
 
 ## 0.6.0.1 -- 2023-11-16